+++ packages/web/src/index.ts
@@ -1,4 +1,4 @@
-import { init, log, SDK_NAME, SDK_VERSION } from 'monitor-browser'
-import { MonitorVue } from 'monitor-vue'
+import { init, log, SDK_NAME, SDK_VERSION } from '@null/monitor-browser'
+import { MonitorVue } from '@null/monitor-vue'
 
 export { init, log, MonitorVue, SDK_NAME, SDK_VERSION }