+++ packages/core/CHANGELOG.md
@@ -9,6 +9,6 @@
 ### Patch Changes
 
 - Updated dependencies
-  - monitor-shared@0.1.0
-  - monitor-types@0.1.0
-  - monitor-utils@0.1.0
+  - @null/monitor-shared@0.1.0
+  - @null/monitor-types@0.1.0
+  - @null/monitor-utils@0.1.0