+++ packages/web-performance/CHANGELOG.md
@@ -5,5 +5,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  - monitor-browser@0.1.0
-  - monitor-vue@0.1.0
+  - @null/monitor-browser@0.1.0
+  - @null/monitor-vue@0.1.0